Rewrote test cases to do content checking.
This actually showed a weird quirk. Technically this is exactly
what my markup lexer is supposed to do, and the output is nonetheless
correct. However I find it odd the output splits on special
characters (Although I know that's what's supposed to happen). It
doesn't really matter if I fix it or not though...