Fixed scaling issue when rendering
I didn't realize this at the time, but I was always rendering in terms
of the virtual resolution. For whatever reason I always made the render
text functions return the "real pixel" width, even though the renderer
is technically working in a virtual resolution.

I have no idea how I never realized this until this morning. Doesn't
matter since the zoom feature is technically not meant to be used (it's
not very practical)