Drag file functionality works.
As it's pretty evident, I've given up any notion of code quality. I
do actually use this tool, so if it works I'm letting it pass, since
I'm not rewriting this in any way shape or form.

It will crash if you attempt to read anything that is not a
slide. That's fine since you are expected to not do that, and
considering this thing boots in the blink of an eye, and does no file
modification or streaming (or otherwise anything that keeps file
handles open), there's no way I can corrupt any slide files or
whatever.

TL;DR the crash is a pretty insignificant cost since it does no damage
at all. It takes less than a second to boot up so I'm willing to
excuse it.