Fault tolerance for missing images.
This program only loads from it's current surrounding path, so it's
pretty easy to have resources that just don't load.

Which is kind of bad, but whatever... I could just make the program set
it's "working" directory to be in the slide path but as the only user my
solution... Is just load the slide in the same directory as the
program lol.